% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis.R
\name{create_md}
\alias{create_md}
\title{Create Markdown File}
\usage{
create_md(analysis, output_name = "Analysis")
}
\arguments{
\item{analysis}{Character string representing the content of the analysis to be written to the markdown file.}

\item{output_name}{Character string representing the base name of the output file and folder. Default is "Analysis".}
}
\value{
None. The function writes the analysis content to a new markdown file.
}
\description{
Creates a markdown file with the specified analysis content, cleans the headers, and appends additional text.
}
