% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_visualization.R
\name{openAIScientist_generate_visualization_rmd}
\alias{openAIScientist_generate_visualization_rmd}
\title{Generate a Visualization Recommendation and ggplot Code in RMarkdown}
\usage{
openAIScientist_generate_visualization_rmd(
  data,
  api_key,
  output_name = "Visualization",
  additional_prompt = ""
)
}
\arguments{
\item{data}{A data frame containing the dataset to analyze.}

\item{api_key}{Your OpenAI API key as a string.}

\item{output_name}{The name of the output file for the R code with ggplot.}

\item{additional_prompt}{Additional instructions for the OpenAI API.}
}
\value{
The generated R code as a string.
}
\description{
This function generates a visualization recommendation and the corresponding ggplot code
for the given dataset using OpenAI's API. The generated R code with the ggplot is saved in a folder as an .Rmd file.
}
\examples{
\dontrun{
data <- data.frame(
  var1 = rnorm(100),
  var2 = rnorm(100),
  outcome = sample(c(0, 1), 100, replace = TRUE)
)
api_key <- "your_openai_api_key"
openAIScientist_generate_visualization_rmd(data, api_key, "Visualization")
}
}
