% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_further_analysis.R
\name{suggest_further_analysis}
\alias{suggest_further_analysis}
\title{Suggest Further Analysis using OpenAI API}
\usage{
suggest_further_analysis(data, api_key)
}
\arguments{
\item{data}{A summary or description of the data.}

\item{api_key}{A character string with the OpenAI API key.}
}
\value{
A character string with suggestions for further analysis.
}
\description{
This function suggests further analyses based on the provided data using the OpenAI API.
}
\examples{
suggest_further_analysis("The data consists of measurements of various plant species.", "your_openai_api_key")
}
